Refactored API files
